lot's of replacing `long` with `int` for many files

+++ MyHebrewBible/MyHebrewBible.Client/Components/Report/Detail/WordSelected.cs
@@ -1,5 +1,5 @@
 ﻿namespace MyHebrewBible.Client.Components.Report.Detail;
 
-public record WordSelected(long Strongs, long WordCount);
+public record WordSelected(int Strongs, int WordCount);
 
 // Ignore Spelling: Strongs
