+++ MyHebrewBible/MyHebrewBible.Client/Features/Parasha/Enums/VerseRange.cs
@@ -3,5 +3,5 @@
 namespace MyHebrewBible.Client.Features.Parasha.Enums;
 
 // See 057-Strongs-Frequency-Analysis\Notes.md re.  `VerseRange.cs` ! `GetSatVerseList()`
-public record VerseRange(BibleBook BibleBook, string ChapterVerse, long	BegId, long EndId);
+public record VerseRange(BibleBook BibleBook, string ChapterVerse, int	BegId, int EndId);
